// powerapps-project-planner-20260612/src/generated/models/Vibe_resourceallocationsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Vibe_resourceallocationsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Vibe_resourceallocationsstatecode = keyof typeof Vibe_resourceallocationsstatecode;
export const Vibe_resourceallocationsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Vibe_resourceallocationsstatuscode = keyof typeof Vibe_resourceallocationsstatuscode;

export interface Vibe_resourceallocationsBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Vibe_resourceallocationsstatecode;
  statuscode?: Vibe_resourceallocationsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
  vibe_actualhours?: number;
  vibe_name?: string;
  vibe_notes?: string;
  vibe_plannedhours?: number;
  "vibe_projectid@odata.bind"?: string;
  "vibe_projectteammemberid@odata.bind"?: string;
  vibe_resourceallocationid: string;
  vibe_weekstartdate?: string;
}

export interface Vibe_resourceallocations extends Vibe_resourceallocationsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  vibe_projectidname?: string;
  vibe_projectteammemberidname?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  vibe_projectid?: object;
  _vibe_projectid_value?: string;
  vibe_projectteammemberid?: object;
  _vibe_projectteammemberid_value?: string;
}